test: hasSpecialEffectのテストをmove-utils.spec.tsに移行
check-registry-coverage.spec.ts内のインラインhasSpecialEffectテストを
move-utils.spec.tsに分離し、責務を明確化

Made-with: Cursor
